refactor(报警参数): 将报警参数类型从uint32改为float以支持更精确的数值
修改报警参数相关函数和结构体，使用float类型替代uint32
更新Modbus从机处理逻辑以支持float类型报警值
调整UI界面以适配float类型的报警参数显示和设置

+++ App/app_config.h
@@ -116,23 +116,23 @@ void app_config_set_modbus_stopbits(uint32_t value);
 /*============================================================================*/
 /*                           报警参数 Getter/Setter                             */
 /*============================================================================*/
-uint32_t app_config_get_alarm_ah(void);
-void app_config_set_alarm_ah(uint32_t value);
+float app_config_get_alarm_ah(void);
+void app_config_set_alarm_ah(float value);
 
-uint32_t app_config_get_alarm_al(void);
-void app_config_set_alarm_al(uint32_t value);
+float app_config_get_alarm_al(void);
+void app_config_set_alarm_al(float value);
 
-uint32_t app_config_get_alarm_dh(void);
-void app_config_set_alarm_dh(uint32_t value);
+float app_config_get_alarm_dh(void);
+void app_config_set_alarm_dh(float value);
 
-uint32_t app_config_get_alarm_dl(void);
-void app_config_set_alarm_dl(uint32_t value);
+float app_config_get_alarm_dl(void);
+void app_config_set_alarm_dl(float value);
 
-uint32_t app_config_get_alarm_aah(void);
-void app_config_set_alarm_aah(uint32_t value);
+float app_config_get_alarm_aah(void);
+void app_config_set_alarm_aah(float value);
 
-uint32_t app_config_get_alarm_aal(void);
-void app_config_set_alarm_aal(uint32_t value);
+float app_config_get_alarm_aal(void);
+void app_config_set_alarm_aal(float value);
 
 /*============================================================================*/
 /*                           其他参数 Getter/Setter                             */
